Remove `ref` restriction to allow `StringSegment` being a field of class.
`StringSegment` has no byref like fields(like a managed pointer pointed to stack) that prevents it to be stored on managed heap. Instead, it stores a string reference, therefore this struct is more similar to `System.ReadOnlyMemory<char>` instead of the Span brother. 
We can make it a regular struct to have more capabilities.